+++ src/shared/property.js
